requirements.txt
langchain
langgraph~=1.0.2
langchain-google-genai~=3.0.0
langsmith
fastapi~=0.116.1
uvicorn~=0.35.0
python-dotenv~=1.2.1
faiss-cpu
tiktoken
langchain-community~=0.4.1
langchain-text-splitters~=1.0.0
sqlalchemy~=2.0.44
langchain-core~=1.0.2
pydantic~=2.12.3
aiosqlite~=0.21.0
pandas~=2.3.3
mcp~=1.21.0
fastapi
uvicorn
langchain
langgraph
langchain-groq
langchain-google-genai
langchain-community
langchain-mcp-adapters
aiosqlite
pandas
openpyxl
python-dotenv
replicate
tavily-python
faiss-cpu
sentence-transformers
requests
gradio_client
cloudinary